Add OpenVINO initialization routine
While the README clearly states how to enable OpenVINO the code
is missing the call to load the OpenVINO encoder. This change
adds the msising call.

Note that the compute device is currently hard-coded. I suspect
that OpenVINO is mainly used for ARM architecture where it doesn't
matter. If need arises the compute device parameter can be added
at a later point in time.